requirements.txt
pdf2docx
pdf2image
pytesseract
python-docx
Pillow
PyPDF2
